requirements.txt
Pillow
Geopy
customtkinter